Modify the camera part to be compatible with espidf v5.5.4, and add RF switch settings in lr2021.

+++ components/private_library/t_display_p4_driver.cpp
@@ -2,7 +2,7 @@
  * @Description: t_display_p4_driver
  * @Author: LILYGO_L
  * @Date: 2025-07-07 14:31:29
- * @LastEditTime: 2025-08-15 15:00:40
+ * @LastEditTime: 2026-06-10 11:29:14
  * @License: GPL 3.0
  */
 #include "esp_lcd_panel_ops.h"
@@ -25,7 +25,7 @@ bool Mipi_Dsi_Init(uint8_t num_data_lanes, uint32_t lane_bit_rate_mbps, uint32_t
         .bus_id = 0,
         .num_data_lanes = num_data_lanes,
         .phy_clk_src = MIPI_DSI_PHY_CLK_SRC_DEFAULT,
-        .lane_bit_rate_mbps = lane_bit_rate_mbps,
+        .lane_bit_rate_mbps = (float)lane_bit_rate_mbps,
     };
 
     esp_err_t assert = esp_lcd_new_dsi_bus(&bus_config, &mipi_dsi_bus);
@@ -51,7 +51,7 @@ bool Mipi_Dsi_Init(uint8_t num_data_lanes, uint32_t lane_bit_rate_mbps, uint32_t
     esp_lcd_dpi_panel_config_t dpi_config = {
         .virtual_channel = 0,
         .dpi_clk_src = MIPI_DSI_DPI_CLK_SRC_DEFAULT,
-        .dpi_clock_freq_mhz = dpi_clock_freq_mhz,
+        .dpi_clock_freq_mhz = (float)dpi_clock_freq_mhz,
         .pixel_format = color_rgb_pixel_format,
         .num_fbs = num_fbs,
         .video_timing = {
